requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
pycodestyle==2.8.0
pycryptodome==3.14.1
pySmartDL==1.3.4
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
